Move Patcher logic to subproject MMR.Patch
