+++ MMR.Randomizer/MMR.Randomizer.csproj
@@ -27,11 +27,11 @@
     <PackageReference Include="SixLabors.ImageSharp" Version="1.0.0-beta0007" />
     <PackageReference Include="SixLabors.ImageSharp.Drawing" Version="1.0.0-beta0007" />
     <PackageReference Include="System.CodeDom" Version="4.7.0" />
-    <PackageReference Include="VCDiff" Version="3.2.7" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\MMR.Common\MMR.Common.csproj" />
+    <ProjectReference Include="..\MMR.Patch\MMR.Patch.csproj" />
     <ProjectReference Include="..\MMR.Rom\MMR.Rom.csproj" />
     <ProjectReference Include="..\MMR.Yaz\MMR.Yaz.csproj" />
   </ItemGroup>
