+++ Majora's Mask Randomiser.sln
@@ -25,7 +25,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MMR.Yaz.CLI", "MMR.Yaz.CLI\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MMR.Yaz.Tests", "MMR.Yaz.Tests\MMR.Yaz.Tests.csproj", "{27B43FBF-4CEB-4154-8EA2-1B959B21630D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MMR.Rom", "MMR.Rom\MMR.Rom.csproj", "{AED904E3-EC31-4898-B0C7-753CBDC10F59}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MMR.Rom", "MMR.Rom\MMR.Rom.csproj", "{AED904E3-EC31-4898-B0C7-753CBDC10F59}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MMR.Patch", "MMR.Patch\MMR.Patch.csproj", "{54A9B6E7-69DC-45EF-A6C2-BEEC618EA57F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -131,6 +133,14 @@ Global
 		{AED904E3-EC31-4898-B0C7-753CBDC10F59}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AED904E3-EC31-4898-B0C7-753CBDC10F59}.Release|x86.ActiveCfg = Release|Any CPU
 		{AED904E3-EC31-4898-B0C7-753CBDC10F59}.Release|x86.Build.0 = Release|Any CPU
+		{54A9B6E7-69DC-45EF-A6C2-BEEC618EA57F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{54A9B6E7-69DC-45EF-A6C2-BEEC618EA57F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{54A9B6E7-69DC-45EF-A6C2-BEEC618EA57F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{54A9B6E7-69DC-45EF-A6C2-BEEC618EA57F}.Debug|x86.Build.0 = Debug|Any CPU
+		{54A9B6E7-69DC-45EF-A6C2-BEEC618EA57F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{54A9B6E7-69DC-45EF-A6C2-BEEC618EA57F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{54A9B6E7-69DC-45EF-A6C2-BEEC618EA57F}.Release|x86.ActiveCfg = Release|Any CPU
+		{54A9B6E7-69DC-45EF-A6C2-BEEC618EA57F}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
